backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
jinja2==3.1.4
google-cloud-storage==2.18.2
google-cloud-firestore==2.17.2
google-auth==2.35.0
pytest==8.3.3
pytest-asyncio==0.24.0
